+++ lib/editor/controllers/vertical-caret-movement-run.controller.dart
@@ -4,8 +4,7 @@ import 'package:flutter/rendering.dart';
 
 import '../widgets/editor-renderer-inner.dart';
 
-class VerticalCaretMovementRunController
-    extends BidirectionalIterator<TextPosition> {
+class VerticalCaretMovementRunController {
   VerticalCaretMovementRunController(
     this._renderer,
     this._currentTextPosition,
@@ -15,18 +14,15 @@ class VerticalCaretMovementRunController
 
   final EditorRendererInner _renderer;
 
-  @override
   TextPosition get current {
     return _currentTextPosition;
   }
 
-  @override
   bool moveNext() {
     _currentTextPosition = _renderer.getTextPositionBelow(_currentTextPosition);
     return true;
   }
 
-  @override
   bool movePrevious() {
     _currentTextPosition = _renderer.getTextPositionAbove(_currentTextPosition);
     return true;